Big Sprint, fixed interface and run.py

+++ Core/Modify.py
@@ -10,23 +10,27 @@ __author__ = 'Joe Bruno'
 
 def main(fileList = None, isModule = True):
     # To be the list of all found files in need of modification
+    li = fileList
+    print "Items found: %d"% len(li)
+    for i, item in enumerate(li):
+        print "%s: %s" % (i+1, item)
     if not fileList and isModule:
         raise AttributeError('No Files were passed in!')
 
-    # variables used to help with testing - can keep hitting run without any damage
-    # This portion will probably be moved elsewhere - just quick to do it here
-    rootDir = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
-    sampleDir = "samples"
-    testDir = "test"
-    testFile = "MyRunM001C01.txt"
-    sampleFilePath = os.path.join(rootDir, sampleDir, testFile)
-    testFilePath = os.path.join(rootDir, testDir, testFile)
+    if not isModule:
+        # variables used to help with testing - can keep hitting run without any damage
+        # This portion will probably be moved elsewhere - just quick to do it here
+        rootDir = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
+        sampleDir = "samples"
+        testDir = "test"
+        testFile = "MyRunM001C01.txt"
+        sampleFilePath = os.path.join(rootDir, sampleDir, testFile)
+        testFilePath = os.path.join(rootDir, testDir, testFile)
 
-    shutil.copyfile(sampleFilePath, testFilePath) # copy sample to test directory
-    os.remove(os.path.join(rootDir, testDir, testFile[:testFile.rfind('.')]+'.awd'))
-    print("Testing with File: %s" % testFilePath)
-    fileList = [testFilePath]
-
+        shutil.copyfile(sampleFilePath, testFilePath) # copy sample to test directory
+        os.remove(os.path.join(rootDir, testDir, testFile[:testFile.rfind('.')]+'.awd'))
+        print("Testing with File: %s" % testFilePath)
+        fileList = [testFilePath]
     modFiles(fileList)
 
 def modFiles(fileLocList):
